requirements.txt
arcgis2geojson==2.0.0
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
Django==3.2.16
django-environ==0.10.0
django-extensions==3.2.1
django-leaflet==0.27.1
django-simple-history==2.12.0
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
geographiclib==1.50
geopy==2.0.0
idna==2.10
install==1.3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
prettytable==2.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-Levenshtein==0.12.0
pytz==2021.3
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2023.3
urllib3==1.26.2
uWSGI==2.0.19.1
wcwidth==0.2.5